requirements.txt
hopsworks[python]==4.0.1
hsfs==4.0.1
pandas<2.0.0
python-dotenv==1.0.1
twofish
yfinance==0.2.54